:recycle: [Create Beneficiary] Move secretQuestion fields into a separated form type
