+++ templates/v2/common/_button.html.twig
@@ -1,3 +1,6 @@
+{% set color = color is defined ? color : 'green' %}
+{% set type = type is defined ? type : 'submit' %}
+{% set message = message is defined ? message : 'submit' %}
 <button class="btn btn-{{ color }} text-white text-center mt-2 col mx-1" type="{{ type }}">
     {{ message|trans }}
 </button>
